requirements.txt
asgiref==3.7.2
Django==4.2.7
django-phonenumber-field==7.2.0
Faker==20.1.0
phonenumbers==8.13.26
phonenumberslite==8.13.26
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
